+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Outline: Shrank llms-txt output to <200_000 input tokens
 - Bundle: Added "count-tokens" procedure to inform about the size
   of the outcome. Sonnet and Opus are limited to 200_000 input tokens.
+- Query: Added a few more example questions specific to CrateDB
 
 ## v0.0.7 - 2025-07-22
 - Prompt: Added `instructions-general.md` file when generating bundle